% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clpt.R
\name{cltp}
\alias{cltp}
\title{POISSON}
\usage{
cltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{the sample size}

\item{iter}{how many times to iterate}

\item{lambda}{lambda}
}
\value{
There is no return but through plot output.
}
\description{
POISSON
}
