% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty.R
\name{myboot2}
\alias{myboot2}
\title{POISSON}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{how many times to iterate}

\item{x}{data}

\item{fun}{function to be ran}
}
\value{
There is no return but through plot output.
}
\description{
POISSON
}
