% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Sample Generator}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{sample vector}

\item{alpha}{alpha value}

\item{beta}{beta value}
}
\value{
Returns x amount of random beta values
}
\description{
Sample Generator
}
