% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvywurvy.R
\name{myncurve}
\alias{myncurve}
\title{Sample Generator}
\usage{
myncurve(mu, sigma)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{stdev}
}
\value{
There is no return but through plot output.
}
\description{
Sample Generator
}
