% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{mysample}
\alias{mysample}
\title{Sample Generator}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{the population size}

\item{iter}{the amount of samples to make}

\item{time}{the amount of time to wait before releasing a new sample}
}
\value{
There is no return but through plot output.
}
\description{
Sample Generator
}
