% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Sample Generator}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the population size}

\item{gamma}{Chance of overpopulation}

\item{p}{probability that people will show}
}
\value{
There is no return but through plot output.
}
\description{
Sample Generator
}
